fix: 채팅 버그 수정
